small fixes, remove temp constructor
