+++ mllib/src/main/scala/org/apache/spark/mllib/classification/LogisticRegression.scala
@@ -430,10 +430,9 @@ class LogisticRegressionWithLBFGS
         lr.setElasticNetParam(elasticNetParam)
         lr.setStandardization(useFeatureScaling)
         if (userSuppliedWeights) {
-          // TODO: check this
           val uid = Identifiable.randomUID("logreg-static")
-          lr.setInitialModel(new org.apache.spark.ml.classification.LogisticRegressionModel(
-            uid, new DenseMatrix(1, initialWeights.size, initialWeights.toArray, isTransposed=true),
+          lr.setInitialModel(new org.apache.spark.ml.classification.LogisticRegressionModel(uid,
+            new DenseMatrix(1, initialWeights.size, initialWeights.toArray, isTransposed = true),
             Vectors.dense(0.0).asML, 2, false))
         }
         lr.setFitIntercept(addIntercept)
